backend/requirements.txt
fastapi[standard]
uvicorn
faster-whisper
speechbrain
spacy
scikit-learn
supabase
xgboost
python-multipart
python-dotenv
librosa
pyannote.audio
motor
dnspython
PyJWT
passlib
bcrypt
httpx
